Reloading
Added a new autonomous that shoots once, reloads, then shoots again.
Added a section to the teleop that controls the reload arm.

We'll need to adjust some values once the arm actually goes on, of
course. Also, because auton_shoot_twice pulls back autonomously, we'll
need to start it from the same place each time.